feat(scan): add multi-threading support
Added support for multi-threading in the API scanning process. Introduced a new flag `max` to specify the maximum number of threads to be used. The `apiscan` function now accepts an additional parameter `maxThreads` to control the concurrency. Implemented a semaphore using a channel to limit the number of goroutines executed simultaneously.